+++ Changelog.md
@@ -17,6 +17,7 @@ xxx
 ### Changed
 - when rescan the tag a new random playlist is generated when the player is closed
 - added option in folder to play new file on rescan
+- grid view audio files
 
 ## [0.17.0]
 ### Added